Refactor main classes and index into Client
