+++ packages/maps/__tests__/Client.spec.ts
@@ -1,21 +1,21 @@
-import Maps from '../src';
+import { Maps as Client} from '../src/Client';
 
 describe('maps', () => {
   it('can be easily created', () => {
-    const m = new Maps('aUser', 'anApiKey');
+    const m = new Client('aUser', 'anApiKey');
     expect(m).toBeTruthy();
   });
 
   it('requires credentials', () => {
     expect(() => {
       // tslint:disable-next-line: no-unused-expression
-      new Maps('', '');
+      new Client('', '');
     }).toThrow('Username is required');
   });
 
   describe('create a simple map', () => {
     it('fails without dataset or sql query', async () => {
-      const m = new Maps('a-user', 'wadus');
+      const m = new Client('a-user', 'wadus');
       await expect(m.instantiateMapFrom({})).rejects
         .toThrowError('Please provide a dataset or a SQL query');
     });
